Implement failover semantics
When a reconciliation loop is invoked, all the PostgreSQL instances will
be asked for their replication status and, if no master is detected, a
standby will be elected and written in `targetPrimary`. This will allow,
in a later work to be handled by pgk, to actually promote the elected
standby.

The labels in the Pods will be adjusted too (when `currentPrimary` will
be aligned by pgk) to get the `-rw` service back to work.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>