Switchover implementation in PGK
The switchover procedure is entirely handled by PGK and not guided by
the operator.

PGK during its reconciliation cycle will look for changes in the
`targetPrimary` status field of the cluster.

When a new server is promoted as a primary:

- the old primary will request a fast shutdown, leading to a restart in
  the Pod. The Pod will be restarted by the Kubelet, and PGK will wait for
  the switchover to be completed (targetPrimary == currentPrimary). It will
  then demote itself, and start following the new primary.

- the promoted replica will wait for its WAL sender to be stopped and
  then promote itself.

Since `currentPrimary` and `targetPrimary` are stored in the status
subresource of the Cluster, they can't currently be modified the
`kubectl command` as per this bug here:

https://github.com/kubernetes/kubectl/issues/564

In this commit there is also included a new command that can be used to
promote PostgreSQL instances.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Co-authored-by: Francesco Canovai <francesco.canovai@2ndquadrant.it>