Fix switchover E2E test
The switchover E2E suite was using `exec` instead of just looking at Pod
labels because the latter wasn't getting the latest Pod status.

The problem was in the `Get` call, which is designed to work on empty
structure instead of full ones. Because of this, the removal of a label
was not detected by the code since the old label in the Pod structure
wasn't removed.

This patch remove every usage of this technique, creating a new object
every time.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>