Add checks for Pods which are being updated
This patch adds several functions to check if a Pod is being updated, that
is if we are changing the image that a container in the Pod is currently
running.

Together with this condition check, the patch adds a function to control
if two image names are equal considering default prefixes (such as
`docker.io/library`) and suffixes (such as `latest`).

All this changes are meant to prepare the ground for a successive commit
regarding rolling update management for PostgreSQL versions.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>