Handle rolling update
If the image required by the user changes, the operator will now start
rolling the upgrade to replicas.

After all the replicas are updated the master need to be updated, and
for that the user can choose between two behaviors, which are
specificable in Spec of the cluster:

- `switchover`: promote the replica with less data to receive or to
  reapply as the new master and then finish the upgrade;

- `wait`: wait for the user to require a switchover, and then finish the
  rolling update process.

In the cluster status we keep track of when we required a Pod to restart
to be updated, and use that one to check if the Pod have been really
updated or not.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>