Reuse existing storage for Pods
If a storage area already exists but the Pod does not, the latter we be
recreated using the old storage area. This is useful for rolling
upgrades and to make maintenance operation easier.

As an example, if a node where a PostgreSQL instance is running is
drained, the Pod is created with the same identity in another node.

This powers the rolling update mechanics, which will simply delete the
Pod to be upgraded waiting for them to be recreated.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>