Allow using community PostgreSQL images
In this moment the operator required to be used with a set of customized
images for PostgreSQL as the controller is supposed available in the
image.

With this patch the controller has been directly included into the
operator and the operator itself, used as an init container for
PostgreSQL instances, will copy his executable in an emptyDir volume to
be used by the following containers.

This will allow also to reuse the mechanisms we already have to manage
the rolling update procedure to manage updates of the operator and of
the controller.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>