Add Backup and WAL streaming support
This patch adds a configuration item to enable WAL streaming for the
PostgreSQL cluster, which is based on 'barman-cloud-wal-archive'.

During the bootstrap, PostgreSQL's 'archive_mode' is always set to 'on',
and 'archive_command' is set to archive the WALs via the manager which
is inside the Pod. WALs are only archived in the master server.

A new custom resource, named Backup, have been introduced. When the
operator spot that a resource like that is created, it will:

- execute 'manager backup <backup name>' in the Pod of the master
  instance;

- the command will issue a REST request to the instance manager
  requesting a new backup;

- the instance manager will start 'barman-cloud-backup' with the
  required parameters.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>