Add scheduled backup support
Add a new CRD, named `ScheduledBackup`, which is similar to `Backup` but
has a `schedule` field added. This field is a cron-like job schedule
with a seconds fields prepended.

A new controller will take care of creating a backup object when the
schedule is met.

From this patch CNP is dependent on the "github.com/robfig/cron"
package, which is the one also used by k8s.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>