Node maintenanance support for PostgreSQL clusters with local storage
When Pods were evicted, the operator normally recreates them and the scheduler
choses another available node. While this is perfectly fine in a shared
storage architecture, it is a Catch22 scenario if local storage is used.
The PV is tied to node and the new Pod cannot be scheduled elsewhere,
impacting also PDB.

The patch introduces the `nodeMaintenanceWindow` parameter, allowing
the user to temporarily place the cluster in maintenance mode
and for example decide to either wait for the node to come up again
(`reusePVC = on`) or migrate the data on another node (`reusePVC = off`).

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@2ndquadrant.it>