Always reuse existing PVCs, even when the cluster is empty
Before this patch the operator was wrongly generating a new master when the
cluster was empty, even if there were reusable PVCs. Now it's possible to
recreate a cluster given the PVCs.

This patch also fix an error inside the instance controller, which was not
executing a control loop the first time an instance was run.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>