Remove `emptyDir` volume support
The operator was using emptyDir storage if the user didn't specify a
StorageClass to be used. Given that the `emptyDir` storage is ephimeral and
meant to be used only when bootstrapping a development environment, we decided
that this wasn't a safe default.

We now require the user to specify a size requirement and, if the StorageClass
is empty, we let k8s use the default storage class of the cluster.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>