Make PostgreSQL configuration optional
The operator now supplies default for the PostgreSQL configuration
parameters and enforce that the required parameters are actually used.

This also lays the foundation for updating the PostgreSQL configuration
via the operator, as the PostgreSQL configuration managed by the
operator is now stored in a separate file and applied every time a Pod
starts up.

This now also happens for `pg_hba.conf`, with the same lifecycle.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@2ndquadrant.it>