Add `bootstrap` configuration section
Before this patch, the operator would create a new PostgreSQL cluster with
`initdb`, and create an application database with an owning user. Now we have a
new configuration section named `bootstrap` which lets the user choose how
to initialize the cluster.

We still support only `initdb` based initialization, but now we let the user
specify if a database needs to be created, and only specify secrets to be used
for the process.

With this patch you can generate a cluster with a specified superuser (and an
app user) password instead of relying on auto-generated ones.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>