Add `fullRecovery` bootstrap type
This new bootstrap type allows the user to recover an existing backup, pointing
the new cluster to the Kubernetes backup object that must be restored.

During the init phase, the primary Pod will:

1. Restore the backup in the data directory using `barman-cloud-restore`
2. Overwrite the configuration with one allowing `trust` access from localhost
   and a `restore_command` to recover files from the object storage
3. Temporarily start the server and wait for it to reach a consistent state
4. Set the new superUser password

As it happens with the `initDb` bootstrap method, a secret can be specified by
the user, allowing it to initialize the database with a certain password.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>