Rename WebHookEnvironment to PublicKeyInfrastructure
Since we are going to support OpenShift, and OLM is generating WebHook
certificates for us, we'll use the infrastructure we have to generate only the
CA for the operator, which will be used to sign the CA we'll create for every
cluster.

Given that, it's no more a `WebHookEnvironment` but a `PublicKeyInfrastructure`
that will be used to generate WebHook certificates, too.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>