Support webhooks certificates generated by OLM
In OpenShift and generally where the operator installation is managed by OLM
the webhooks certificates are already generated in a specified folder.

This patch makes the webhook certificate generation optional allowing the user
to specify an environment variable where the certificate files will be found.
This allows the operator to be installed in OpenShift with a simple patch
adding an environment variable.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>