Filter out reconcile events
This patch add a Predicate filtering events that enqueue a new reconciliation
loop for Clusters. Only events related the specification of the clusters and
the Pod owned by a clusters are now processed.

Before this patch every event related to a Pod in the k8s server could lead to
a reconciliation loop.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>