Apply configuration changes to instances
Now instances are grabbing a watch over the ConfigMap which is created and
updated by the operator and when a configuration change happens they will react
by updating their configuration files. In the instance status a new field named
`pendingRestart` has been added.

This new field will be used by the operator and by the instances to coordinate
the configuration application for parameters which need a restart to be
applied.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>