Use unix domain socket to control the local instance
The instance manager now is using the local socket to control the instance,
instead of opening a TCP connection to 127.0.0.1. Generating the `.pgpass` file
is still needed since we are using it for replicas and for a possibly demoted
primary server.

We are also directly generating a `pg_hba.conf` and `postgesql.conf` file from
the ConfigMap, without using any mountpoints.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Francesco Canovai <francesco.canovai@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>