Enable TLS communication to Postgres instances
The operator now creates a CA for every cluster, signed with the private key of
the operator CA. From that CA we also create a leaf certificate to be used by
the PostgreSQL instances.

The patch also handles the certificate renewal and the injection inside the
actual PostgreSQL instance. After the certificate renewal, PostgreSQL is
reloaded. To do that, an additional watch is created by the Instance Manager
over the server TLS secret.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>