Use `streaming_replica` user for replication
The operator now uses a user named `streaming_replica` for replication instead
of relying on the `postgres` user to do that.

That user is created while bootstrapping a new cluster and after having
restored a physical backup - if needed.

This feature improves the security stance of clusters created by the operator.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>