Support custom `initdb` options
A new option named `options` has been added in the `initdb` bootstrap
configuration section, allowing the user to pass custom option to the database
creation process. This, in turn, allows the user to change some options that
can be set only during the bootstrap phase, like the default encoding for new
databases.

The patch also fix an error during the bootstrap that was preventing the
operator to correctly set the password for the application user.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>