`pg_rewind` needs superuser permissions on PostgreSQL 10
On PostgreSQL 10 individual execution permissions on the function invoked by
`pg_rewind` are not sufficient, you need full superuser permissions.

Besides that, this patch fixes a race condition happening when the termination
of old-primary nodes (to be demoted) happens before the code handling the instance
configuration.

The instance configuration has been moved inside the reconciliation loop while
handling the `ADDED` events generated for the cluster itself.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>