Add PodTemplate to cluster specificaton
This patch adds a new parameter to the CRD named `podTemplate`, containing a
template that is used to create Pods for PostgreSQL instances. This feature is
useful to fine-grain configure things that don't have a direct field in
CRD, like NodeAffinity.

The patch also updates the version of `controllen-gen` we are using the latest version
because the previous one had a bug preventing the generation of CRDs where PodTemplate
is used. More information can be found here:

https://github.com/operator-framework/operator-sdk/issues/3235

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@2ndquadrant.it>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@2ndquadrant.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@2ndquadrant.it>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@2ndquadrant.it>