Fix failover E2E test
The test was pausing replication on an instance, expecting the other one
to be promoted. However, the walreceiver was still receiving WALs, so
the paused instance could still be promoted, making the test fail. We
now send a SIGSTOP to the walreceiver, so we know for sure that nothing
gets on that standby and we can expect the other one to be promoted.

Terminate the pausedReplica walsender on the primary after pausing the
PausedReplica walreceiver. This way, we don't wait for the replication
timeout to occur.

Co-authored-by: Niccolò Fei <niccolo.fei@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>