Remove completed Jobs from the cluster
There's no point in preserving Jobs which are terminated, and now the
operator will remove them when they are completed.

The operator now requeue a reconciliation loop after having detected an
expectation which is not met. In this way, even if the API server
doesn't notify us of something changed, the operator will still be able
to continue its work.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>