Avoid statically settings AWS environment variables in Pods
We were using two Env fields inside the Pods of the instances to inject
the AWS credentials when needed. Doing that we needed to delete the Pods
when adding/changing the backup section inside a cluster.

Instead of doing that we are now getting the actual secrets in the
`wal-archive` and `wal-restore` command, getting the latest settings of
the cluster.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>