Avoid classifying new PVCs as dangling
Before this patch, when a PVC initialization Job completes, the PVC is
classified as dangling and the operator generates a new Pod to use it.

This behavior conflicts with the maintenance window management
when the reuse of PVC is disabled. This patch introduces
a new "initializing" PVC status to avoid classifying the initialized PVC
as dangling and to continue the setup of the new instance.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>