Avoid breaking the reconciliation loop on exec failures
We are now not interrupting the reconciliation cycle when a certain Pod
cannot be reached by the Exec call. When a Pod is not reachable, it is
not electable as a possible primary node: this is not a new behavior as
it was happening also before that patch.

This patch also solves a problem in the `kubectl cnp status` command,
which was not showing the status of non-ready Pods.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>