Add support for annotations inheritance
This commit introduces the new `INHERITED_ANNOTATIONS` setting
that accepts a comma-separated list of annotation names.

The operator will copy annotations from the `Cluster` resource to
the generated resources if their name is in the inherited annotation list.

You can set the `INHERITED_ANNOTATIONS`  setting in the operator
environment or defining a key in the configuration ConfigMap.

The configuration package has been refactored to ease unit testing and
now the configuration is shown in the logs during the operator startup.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>