Remove ConfigMap storing PostgreSQL configuration
Now the instance manager autonomously generates the required PostgreSQL
configuration for the cluster it belongs to. This also removes one
Watch per Pod, lowering the workload of the API server.

This commit removes also a great part of the dynamic Kubernetes
interface from the instance manager, relying on a manual
conversion/deconversion to/from unstructured objects.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>