Avoid watching secrets from the Instance Manager
The instance manager will now avoid watching secrets, as the operator
will change the cluster status every time the secrets are changed.

To do that a new section in the Cluster Status has been added, and this
new section is tracking the ResourceVersion of each generated secrets.

The operator will react changing that resource version when the
corresponding secret is changed, keeping a watch for every generated
Secret. This patch is also the wireframe for a next commit which will
let the user change the 'postgres' password by changing the contents of
the secret.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>