Start postgres only when `pg_rewind` fails
`pg_rewind` could require a clean shutdown of the old primary to
work. Unfortunately, if the old primary is already clean, starting
it up may make it advance in respect to the new one.
The only way to check if we really need to start it up before
invoking `pg_rewind` is to try using `pg_rewind` and, on failures,
retrying after having started up the instance. This patch adds this
step.