Use application database for monitoring queries
If the cluster is created using the `initdb` instance method, we'll now
use the application database for custom monitoring queries. A future
commit will improve the custom quieries API to let the user choose the
database to be used.

The instance manager is now using a connection pooling implementation
allowing it to use different databases instead of just the `postgres`
and the application database one.

Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>