Implement `kubectl cnp restart` command
This command requires the operator to rollout restart a certain cluster
using the the mechanism used by the `kubectl rollout restart` command.

An annotation will be set on the cluster, containing the required restart
time. The value of this annotation will be then used to check is a Pod
need to be recreated or not.

Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>