Prefer controller-runtime over dynamic client in the instance manager
Instead of using the client-go static and dynamic interface in the
instance manager, we are now relying on the controller-runtime client
whenever possible.

The only point where we retain the dynamic interface is in the
controller loop of the instance manager, and we're doing it because
controller-runtime in its current version have no generic interface over
watches.

As soon as we upgrade to controller-runtime 0.9.0 we can remove that
usage too.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>