Allow bootstrapping using `pg_basebackup`
This patch adds a new feature to clone an existing PostgreSQL cluster,
running in a Kubernetes cluster or external. This allows the user to
create a new PostgreSQL cluster from an existent PostgreSQL database
using the replication protocol and the pg_basebackup tool.

In the Cluster API, we introduced a new section named `externalClusters`,
containing the connection parameters to more than one external
PostgreSQL server. Each server has its own name, and that name is then
referenced inside the new bootstrap method.

Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>