Enforce streaming replication connection in ClonePgData method
When cloning an existing instance (creating a replica or creating
a new cluster using the `pg_basebackup` bootstrap method), enforce
a streaming replication connection by adding "replication=1" to the
connection string. As a result, connecting to an external cluster
for `pg_basebackup` bootstrap, does not require to specify any database
name.

Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>