Allow specifying sha256 in the cluster image name
Image `sha256` can be optionally specified in addition to the image tag.
This allows forcing a specific image to be used for a cluster.

The tag is mandatory, as we need the PostgreSQL version to be able to verify
the compatibility.

Switching from the default image / an image without the digest to one
including the sha256 digest triggers a rolling restart of all instances,
even if the deployed image already had the correct one.
The operator compares the two strings from the cluster's and pods' specs.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>