fix: configuration parser boolean defaults
The current configuration parser will emit an error when there is no
configmap nor environment because the boolean default values are parsed
incorrectly.

This doesn't really affect nor change the meaning inside the operator,
since an unparsable default is taken as "false" and this is exactly the
zero value for booleans.