feat: distinct PDBs for primary and replicas
Introducing a dedicated `PodDisruptionBudget` for primary instances,
requiring at least a primary instance to be running at each time. This
allows the operator to better control how instances are handled during a
drain.

Now, whenever the primary instance is running on an unschedulable node
the operator will trigger a switchover to an instance on a schedulable
and healthier node.

Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>