fix: avoid beginning a configuration change from the primary
Given that a new configuration rollout should imply a single
switchover, a race condition between multiple instances updating the
parent cluster resource could lead to the primary instance being
restarted immediately, triggering a switchover to an instance which
still have to be restarted.  For this reason the primary instance, if
other instances are available won't retrigger a reconciliation loop
that could lead to a rolling restart of all the instances, but instead
will wait for the other instances to update the parent cluster.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>