feat: exposing manual switchover required metric
To be able to configure an alert, the primary instance now
exposes the metric cnp_collector_manual_switchover_required. It
will be set to 1 when the cluster's primary requires a manual
switchover to finish an update because "primaryUpdateStrategy" is
set to "supervised" (e.g., because some configuration change
requires a restart).

Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>