fix: allow the operator to failover when the primary is not ready
This patch makes the operator failover to a new Pod when the current
primary is not ready, in accordance with the Pod sorting algorithm that
is already implemented.

This was not the case with the old implementation of the
`updateTargetPrimaryFromPods` function, where the primary Pod (for
primary clusters) was defined as:

1. the active Pod
2. with the same name as the current and target primary
3. if the previous two conditions do not select any Pod, the first of
   the sorted list

This original definition of `primaryPod` was used to elect the new
primary and to check if the Pod is scheduled on a cordoned node.

Now this definition of `primaryPod` will be used only when checking
if the node where the primary Pod is scheduled is cordoned or not, and
not to elect the new primary.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>