feat: allow cluster recovery from an external cluster
This patch adds a new section to each external cluster allowing the user
to specify a backup configuration. A new `source` property has been
added to the recovery bootstrap method. The user can now recover
from a backup that has been taken by a specific external cluster with
a definition like the following excerpt.

```
apiVersion: postgresql.k8s.enterprisedb.io/v1
kind: Cluster
metadata:
  name: cluster-restore
spec:
  instances: 3
  [...]

  bootstrap:
    recovery:
      source: postgresql-bkp

  externalClusters:
    - name: postgresql-bkp
      barmanObjectStore:
        destinationPath: s3://backups/
        [...]
```

The `recover` command of the instance manager has been changed to avoid
relying on the environment variables set inside the Pod and manually
loading the required secrets directly. This behavior is coherent with the other
commands of the instance manager, and allows also to use more than one
object store (the backup object store and the recover object store).

The code only implements the recovery from the latest consistent state
but this lays the fundaments for:

- recovering to a given Point-In-Time
- bootstrapping a replica cluster given an external cluster
- asynchronous replica clusters, recovering WALs of the origin cluster
  from the defined object store

Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Jitendra Wadle <jitendra.wadle@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>