feat: allow non-streaming replica clusters
This patch allows the instance manager to configure a replica cluster
which works without being directly connected to the parent cluster.

The only thing that the replica cluster needs is read access to the
recovery object storage: it will be used to restore the latest physical
backup and to continue applying the WALs archived by the source.

This allows creating a replica cluster which doesn't pose any
traffic/workload on its source, relying on the `archive_timeout`
setting on the source to establish the RPO.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>