feat: add `enableSuperuserAccess` field
When `enableSuperuserAccess` is true, the operator will use the `SuperuserSecret`
to update the `postgres` user password (if the secret is not present, the operator
will automatically create one). This is the default.

When `enableSuperuserAccess` is false, the operator will ignore the `SuperuserSecret`
content, delete it when automatically created, and then blank the password
of the `postgres` user by setting it to `NULL`.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>