feat: set `archive_mode` to `always` on a designated primary
Every time we have a designated primary in a replica cluster, we make
sure that the `archive_mode` is set to `always` to enable WAL shipping
to the local backup object store. If the replica was a previous primary
we make sure that the option `archive_mode` is set to `on`, which is the default.

Co-authored-by: Jihyuk Bok <jihyuk.bok@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>