feat: cache cluster object and subprocess envs in instance manager
With previous implementation every time a subprocess is spawned it was
retrieving the cluster and all the other required objects from the API server,
putting an unnecessary load on the Kubernetes system.

Now the instance manager caches the required objects in memory,
and passes them to the subprocess when needed.

Co-authored-by: Jihyuk Bok <jihyuk.bok@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>