feat: online upgrades of the instance manager
Add a new mechanism that allows the operator to replace the instance
manager in running instances without restarting Postgres.

During online upgrades, the operator uses a new HTTP endpoint in the
manager API to inject a new instance manager executable. The instance
manager verifies the integrity of the injected binary by comparing its
hashcode with the one declared in the Cluster object. Then it gracefully
terminates all the internal processes and replaces itself with the new
instance manager.

When the instance manager restarts using the new binary, it adopts the
already running postmaster. The Postgres process is unaffected by the
manager replacement, so that upgrading the operator does not require any
switchover.

This implementation only works when the operator Pod platform matches
the platform of all instances Pods.

This feature is disabled by default.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Jihyuk Bok <jihyuk.bok@enterprisedb.com>
Co-authored-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>
Co-authored-by: John Long <john.long@enterprisedb.com>
Co-authored-by: Francesco Canovai <francesco.canovai@enterprisedb.com>