feat: handle PostgreSQL shutdown from the instance manager
Now the postmaster termination behavior is handled directly by the
instance manager, avoiding having a lifecycle hook handling that.

The user can choose the maximum time needed to shutdown a PostgreSQL
instance in the cluster specification.
The first half of that time will be spent waiting connections to go down
(smart mode) and, if needed, the second half will be spent to shutting
down the instance itself (fast mode).

Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>