fix: disable PKI infrastructure initialization on OpenShift
On platforms that use OLM, the operator uses the certificates provided
during the installation, so it doesn't need to set up its own PKI
infrastructure.

Also, fix an error in the PKI setup procedure preventing to detect
the operator deployment when the deployment name is different from the
default one.

With this patch, we not only skip initializing the PKI infrastructure
if installed through OLM, but we also delete any eventual leftover from
a previous version.

Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>