fix: liveness probe during PostgreSQL startup
Liveness probe now correctly handles the startup process of a PostgreSQL server.
This is particularly important for standby servers which, during the startup process,
are still in the process of WAL restore and replay to reach a consistent state - and as
such, they cannot accept connections.

This patch fixes some cases of standby servers pods going in `CrashLoopBackOff`
state.

Co-authored-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Co-authored-by: Gabriele Quaresima <gabriele.quaresima@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>