feat: parallel WAL archiving
This patch adds a parallel WAL archiving feature, similar to the one
that Barman has for its archive_command. This new feature can be
enabled via the `maxParallel` option of the WAL archiving configuration.

Every time the archive_command is invoked by PostgreSQL, we pre-archive
a maximum of `maxParallel` WAL files chosen from the ready ones. The
archival status is kept inside a spool folder.

Co-authored-by: Armando Ruocco <armando.ruocco@enterprisedb.com>
Co-authored-by: Philippe Scorsolini <philippe.scorsolini@enterprisedb.com>
Co-authored-by: Marco Nenciarini <marco.nenciarini@enterprisedb.com>
Co-authored-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>